Since the introduction of graphql.js, the graphiql web page has been broken due to a collision in the package name. Renaming graphql.js to graphqlapi.js fixes this issue
